refactor: Refactor teleport request handler and tick loop

+++ src/main/kotlin/net/superricky/tpaplusplus/config/CommonSpec.kt
@@ -5,6 +5,6 @@ import com.uchuhimo.konf.ConfigSpec
 object CommonSpec : ConfigSpec("common") {
     val showBlockedMessage by required<Boolean>()
     val toggledPlayerCommand by required<Boolean>()
-    val tpaTimeout by required<Int>()
-    val waitTimeBeforeTp by required<Int>()
+    val tpaTimeout by required<Double>()
+    val waitTimeBeforeTp by required<Double>()
 }
